refactor: isolated production and dev webpack configuration
