+++ package.json
@@ -6,7 +6,7 @@
   "main": "index.js",
   "scripts": {
     "dev": "webpack-dev-server --env NODE_ENV=development",
-    "build": "webpack --env NODE_ENV=production",
+    "build": "webpack --config webpack-prod.config.js --env NODE_ENV=production",
     "test": "echo \"Error: no test specified\" && exit 1",
     "verify-format": "yarn prettier-check && yarn eclint-check",
     "fix-format": "yarn prettier-fix && yarn eclint-fix",
@@ -32,6 +32,7 @@
     "@babel/preset-react": "^7.16.7",
     "babel": "^6.23.0",
     "babel-loader": "^8.2.3",
+    "copy-webpack-plugin": "^10.2.4",
     "css-loader": "^6.5.1",
     "dotenv": "^16.0.0",
     "dotenv-flow-webpack": "^1.1.0",
